requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langsmith
langgraph
openai
qdrant-client
streamlit==1.28.0
requests==2.31.0
python-dotenv==1.0.0
PyYAML==6.0.3
pypdf2==3.0.1
pytest==7.4.0
sentence-transformers==2.2.2
fastembed==0.7.3
pypdf==6.1.1




